# src/server/swagger/swagger.json
{
  "openapi" : "3.0.2",
  "info" : {
    "title" : "Network Detection and Response (NDR)",
    "description" : "Advanced Cybersecurity Analytics Service (ACAS) for AI-based analysis with explainable intelligence of network events captured by the Montimage Monitoring Tools (MMT)",
    "termsOfService" : "https://github.com/Montimage/acas",
    "contact" : {
      "email" : "developer@montimage.com"
    },
    "license" : {
      "name" : "Montimage",
      "url" : "https://www.montimage.eu"
    },
    "version" : "v1"
  },
...
  "servers": [
    {
      "description": "Localhost Server",
      "url": "http://localhost:31057/api/"
    },
    {
      "description": "Production Server",
      "url": "https://ndr.montimage.eu/api/"
    }
  ],
...
  "paths" : {
...
    "/retrain/predictions/{retrainId}" : {
      "get" : {
        "tags" : [ "models" ],
        "summary" : "Get predictions from retrain",
        "description" : "Get predictions generated during retraining",
        "operationId" : "getRetrainPredictions",
        "parameters" : [ {
          "name" : "retrainId",
          "in" : "path",
          "description" : "Retrain ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "text/csv" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/models/{modelId}/predictions" : {
      "get" : {
        "tags" : [ "models" ],
        "summary" : "Get predictions of the model",
        "description" : "Get predictions of the model",
        "operationId" : "getModelPredictions",
        "parameters" : [ {
          "name" : "modelId",
          "in" : "path",
          "description" : "Name of the model",
          "required" : true,
          "schema" : {
            "type" : "string",
            "example" : "3c6e4cb4-9b74-40c8-8f4a-132c9d5e4ace.h5"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "string",
                  "example" : ""
                }
              }
            }
          }
        }
      }
    },
...
    "/predictions/{predictionId}" : {
      "get" : {
        "tags" : [ "predictions" ],
        "summary" : "Get the statistics of prediction",
        "description" : "Get the statistics of prediction where there are total number of flows, number of attack flows, and number of normal flows",
        "operationId" : "getPredictionStats",
        "parameters" : [ {
          "name" : "predictionId",
          "in" : "path",
          "description" : "Name of the model file to return",
          "required" : true,
          "schema" : {
            "type" : "string",
            "example" : "4bbfaa76-9e39-49b8-baa7-c4a7cdc9fed8"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "prediction" : {
                      "type" : "array",
                      "items" : {
                        "type" : "number",
                        "example" : 1243
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/predictions/{predictionId}/download" : {
      "get" : {
        "tags" : [ "predictions" ],
        "summary" : "Download a prediction file",
        "description" : "Download a prediction file",
        "operationId" : "downloadPrediction",
        "parameters" : [ {
          "name" : "predictionId",
          "in" : "path",
          "description" : "Name of the predictionId to return",
          "required" : true,
          "schema" : {
            "type" : "string",
            "example" : "3c6e4cb4-9b74-40c8-8f4a-132c9d5e4ace"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "multipart/form-data" : {
                "schema" : {
                  "type" : "string",
                  "format" : "binary"
                }
              }
            }
          },
          "401" : {
            "description" : "Prediction does not exist"
          }
        }
      }
    },
    "/predictions/{predictionId}/attack" : {
      "get" : {
        "tags" : [ "predictions" ],
        "summary" : "Download full report on attack traffics",
        "description" : "Download full report of attack traffics",
        "operationId" : "downloadPredictionAttack",
        "parameters" : [ {
          "name" : "predictionId",
          "in" : "path",
          "description" : "Name of the predictionId to return",
          "required" : true,
          "schema" : {
            "type" : "string",
            "example" : "3c6e4cb4-9b74-40c8-8f4a-132c9d5e4ace"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "multipart/form-data" : {
                "schema" : {
                  "type" : "string",
                  "format" : "binary"
                }
              }
            }
          },
          "401" : {
            "description" : "Prediction does not exist"
          }
        }
      }
    },
    "/predictions/{predictionId}/normal" : {
      "get" : {
        "tags" : [ "predictions" ],
        "summary" : "Download full report of normal traffics",
        "description" : "Download full report of normal traffics",
        "operationId" : "downloadPredictionNormal",
        "parameters" : [ {
          "name" : "predictionId",
          "in" : "path",
          "description" : "Name of the predictionId to return",
          "required" : true,
          "schema" : {
            "type" : "string",
            "example" : "3c6e4cb4-9b74-40c8-8f4a-132c9d5e4ace"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "multipart/form-data" : {
                "schema" : {
                  "type" : "string",
                  "format" : "binary"
                }
              }
            }
          },
          "401" : {
            "description" : "Prediction does not exist"
          }
        }
      }
    },
...
  },
...
}